requirements.txt
click
setuptools
pyyaml
